+++ src/visuals-api.d.ts
@@ -1529,6 +1529,27 @@ declare module powerbi.extensibility {
     }
 }
 
+declare module powerbi.extensibility {
+    export interface AcquireAADTokenResult {
+        accessToken?: string;
+    }
+
+    export interface IAcquireAADTokenService {
+        /** Returns an authentication token for the resource that the visual defined as a privilge
+         * and the scope is the visual guid plus a constant string "_CV_ForPBI"
+         * @returns the promise that resolves to the authentication token
+        */
+        acquireAADToken(): IPromise<AcquireAADTokenResult>;
+
+        /**
+         * Returns the availability status of the service.
+         * 
+         * @returns the promise that resolves to privilege status of the service
+         */
+        acquireAADTokenstatus(): IPromise<PrivilegeStatus>;
+    }
+}
+
 declare module powerbi {
     /**
  * Represents a return type for privilege status query methods
@@ -1590,6 +1611,50 @@ declare module powerbi.extensibility {
         /**
          * Deletes data associated with 'key' from local storage.
          * 
+         * @param key - the name of the payload to remove
+         */
+        remove(key: string): void;
+    }
+}
+
+declare module powerbi.extensibility {
+
+    interface StorageV2ResultInfo {
+        success: boolean;
+    }
+
+    /**
+     * Provides an access to local storage for read / write access
+     */
+    interface IVisualLocalStorageV2Service {
+        /**
+         * Returns the availability status of the service.
+         *
+         * @returns the promise that resolves to privilege status of the service
+         */
+        status(): IPromise<PrivilegeStatus>;
+
+        /**
+         * Returns promise that resolves to the data associated with 'key' if it was found or rejects otherwise.
+         *
+         * @param key - the name of the payload to retrieve
+         * @returns the promise that resolves to the data required or rejects if it wasn't found or an error occured.
+         */
+        get(key: string): IPromise<string>;
+
+        /**
+         * Saves the data to local storage. This data can be later be retrieved using the 'key'.
+         * Returns a promise that resolves to StorageV2ResultInfo, or rejects if an error occured.
+         *
+         * @param key - the name of the payload to store
+         * @param data - the payload string to store
+         * @returns the promise resolves to StorageV2ResultInfo, or rejects if an error occured.
+         */
+        set(key: string, data: string): IPromise<StorageV2ResultInfo>;
+
+        /**
+         * Deletes data associated with 'key' from local storage.
+         *
          * @param key - the name of the payload to remove
          */
         remove(key: string): void;
@@ -1732,6 +1797,8 @@ declare module powerbi.extensibility.visual {
         webAccessService: IWebAccessService;
         drill: (args: DrillArgs) => void;
         applyCustomSort: (args: CustomVisualApplyCustomSortArgs) => void;
+        storageV2Service: IVisualLocalStorageV2Service;
+        acquireAADTokenService: IAcquireAADTokenService;
     }
 
     export interface VisualUpdateOptions extends extensibility.VisualUpdateOptions {